backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.0
pymysql==1.1.1
cryptography==43.0.0
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic[email]==2.9.0
pydantic-settings==2.5.0
bcrypt==4.2.0
PyJWT==2.9.0
scikit-learn>=1.6.0
pdfplumber==0.11.0
python-docx==1.1.0
pandas>=2.2.0
reportlab==4.2.0
